+++ HISTORY.md
@@ -4,14 +4,14 @@
 
 ### v0.02 22092023 Init project
 #### TODO
-- [x] define all processor parameters => **all but effect specific parameters mapped**
+- [x] define all processor parameters => **all but effect specific fx parameters mapped**
 - [ ] write get/Set state information for all processor parameters
 - [x] continue definition of EffectComponent & OptionsComponent
-- [x] define SequenceSequencerComponent **to be continued-methods to be added**
+- [x] define SequenceSequencerComponent 
 - [ ] update processor gainpPatterns while modifying patterns on GUI (**crashes**)
 - [ ] cleaner way to pass colortemplates
 #### BUGS
-- [ ] setting the textbutton label color(savePresetButton in options component object) make it crash (why???)
+- [x] setting the textbutton label color(savePresetButton in options component object) make it crash (why???)
 
 ### v0.01 21092023 Init project
 #### TODO